requirements.txt
python-dotenv==1.0.1
requests==2.32.3
fastapi==0.115.12
uvicorn==0.34.2
httpx==0.28.1